+++ Common/src/main/java/net/potionstudios/biomeswevegone/config/ConfigUtils.java
@@ -43,4 +43,10 @@ public class ConfigUtils {
 			throw new RuntimeException(e);
 		}
 	}
+
+	public record CommentValue<T>(String comment, T value) {
+		public static <T> CommentValue<T> of(String comment, T value) {
+			return new CommentValue<>(comment, value);
+		}
+	}
 }