+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/entity/npc/BWGVillagerTrades.java
@@ -10,10 +10,15 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.item.trading.ItemCost;
 import net.minecraft.world.item.trading.MerchantOffer;
 import net.minecraft.world.level.ItemLike;
+import net.minecraft.world.level.block.FlowerBlock;
+import net.minecraft.world.level.block.SaplingBlock;
 import net.potionstudios.biomeswevegone.config.configs.BWGTradesConfig;
 import net.potionstudios.biomeswevegone.world.item.BWGItems;
 import net.potionstudios.biomeswevegone.world.level.block.BWGBlocks;
+import net.potionstudios.biomeswevegone.world.level.block.sand.BWGSandSet;
+import net.potionstudios.biomeswevegone.world.level.block.wood.BWGWood;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -23,30 +28,31 @@ public class BWGVillagerTrades {
     public static final Map<VillagerProfession, Int2ObjectMap<List<MerchantOffer>>> TRADES = new HashMap<>();
 
     public static void makeTrades() {
-        TRADES.put(BWGVillagerProfessions.FORAGER.get(), toIntMap(ImmutableMap.of(
-                1, ImmutableList.of(
-                        createEmeraldForItemsOffer(Items.RED_MUSHROOM, 10, 12, 2),
-                        createEmeraldForItemsOffer(Items.BROWN_MUSHROOM, 10, 12, 2),
-                        createEmeraldForItemsOffer(BWGBlocks.GREEN_MUSHROOM.get(), 10, 12, 2)
-                ),
-                2, ImmutableList.of(
-                        createEmeraldForItemsOffer(BWGBlocks.WOOD_BLEWIT.get(), 8, 12, 3),
-                        createItemsForEmeraldsOffer(BWGItems.WHITE_PUFFBALL_CAP.get(), 4, 5, 4, 2, 0.05f)
-                ),
-                3, ImmutableList.of(
-                        createEmeraldForItemsOffer(BWGItems.WHITE_PUFFBALL_SPORES.get(), 4, 5, 4)
-                ),
-                4, ImmutableList.of(
-                        createItemsForEmeraldsOffer(BWGBlocks.WITCH_HAZEL_BRANCH.get(), 4, 9, 4, 3, 0.05f),
-                        createItemsForEmeraldsOffer(BWGBlocks.WITCH_HAZEL_BLOSSOM.get(), 10, 1, 10, 3, 0.05f)
-                ),
-                5, ImmutableList.of(
-                        createItemsForEmeraldsOffer(BWGBlocks.SHELF_FUNGI.get(), 3, 9, 4, 4, 0.05f),
-                        createEmeraldForItemsOffer(Items.SWEET_BERRIES, 16, 4, 2),
-                        createEmeraldForItemsOffer(BWGItems.BLUEBERRIES.get(), 16, 4, 2)
-                )
-        )));
-        if (!BWGTradesConfig.INSTANCE.get().enableVanillaTradeAdditions()) return;
+        if (BWGTradesConfig.INSTANCE.villagerTrades.allowBWGForagerTrades.value())
+            TRADES.put(BWGVillagerProfessions.FORAGER.get(), toIntMap(ImmutableMap.of(
+                    1, ImmutableList.of(
+                            createEmeraldForItemsOffer(Items.RED_MUSHROOM, 10, 12, 2),
+                            createEmeraldForItemsOffer(Items.BROWN_MUSHROOM, 10, 12, 2),
+                            createEmeraldForItemsOffer(BWGBlocks.GREEN_MUSHROOM.get(), 10, 12, 2)
+                    ),
+                    2, ImmutableList.of(
+                            createEmeraldForItemsOffer(BWGBlocks.WOOD_BLEWIT.get(), 8, 12, 3),
+                            createItemsForEmeraldsOffer(BWGItems.WHITE_PUFFBALL_CAP.get(), 4, 5, 4, 2, 0.05f)
+                    ),
+                    3, ImmutableList.of(
+                            createEmeraldForItemsOffer(BWGItems.WHITE_PUFFBALL_SPORES.get(), 4, 5, 4)
+                    ),
+                    4, ImmutableList.of(
+                            createItemsForEmeraldsOffer(BWGBlocks.WITCH_HAZEL_BRANCH.get(), 4, 9, 4, 3, 0.05f),
+                            createItemsForEmeraldsOffer(BWGBlocks.WITCH_HAZEL_BLOSSOM.get(), 10, 1, 10, 3, 0.05f)
+                    ),
+                    5, ImmutableList.of(
+                            createItemsForEmeraldsOffer(BWGBlocks.SHELF_FUNGI.get(), 3, 9, 4, 4, 0.05f),
+                            createEmeraldForItemsOffer(Items.SWEET_BERRIES, 16, 4, 2),
+                            createEmeraldForItemsOffer(BWGItems.BLUEBERRIES.get(), 16, 4, 2)
+                    )
+            )));
+        if (!BWGTradesConfig.INSTANCE.villagerTrades.enableBWGVanillaProfessionTradeAdditions.value()) return;
         TRADES.put(VillagerProfession.BUTCHER, toIntMap(ImmutableMap.of(
                 2, ImmutableList.of(
                         createEmeraldForItemsOffer(BWGItems.BLUEBERRIES.get(), 10, 12, 2)
@@ -79,6 +85,24 @@ public class BWGVillagerTrades {
                         createItemsForEmeraldsOffer(BWGBlocks.RED_ROCK_SET.getBase(), 1, 1, 12, 15, 0.05f)
                 )
         )));
+    }
+
+    public static final Int2ObjectMap<List<MerchantOffer>> WANDERING_TRADER_TRADES = new Int2ObjectOpenHashMap<>();
+
+    public static void makeWanderingTrades() {
+        List<MerchantOffer> level1Items = new ArrayList<>();
+        BWGWood.WOOD.stream().filter(item -> item.get() instanceof SaplingBlock).forEach(item ->
+                level1Items.add(createItemsForEmeraldsOffer(item.get(), 5, 1, 8, 1, 0.05f)));
+        BWGSandSet.getSandSets().forEach(bwgSandSet -> level1Items.add(
+                createItemsForEmeraldsOffer(bwgSandSet.getSand(), 1, 8, 8, 1, 0.05f)));
+        BWGBlocks.BLOCKS.stream().filter(block -> block.get() instanceof FlowerBlock).forEach(block ->
+                level1Items.add(createItemsForEmeraldsOffer(block.get(), 1, 1, 13, 1, 0.05f)));
+        level1Items.add(createItemsForEmeraldsOffer(BWGItems.TINY_LILY_PADS.get(), 1, 2, 5, 1, 0.05f));
+        level1Items.add(createItemsForEmeraldsOffer(BWGItems.FLOWERING_TINY_LILY_PADS.get(), 1, 2, 5, 1, 0.05f));
+        level1Items.add(createItemsForEmeraldsOffer(BWGBlocks.WEEPING_MILKCAP.get(), 1, 1, 12, 1, 0.05f));
+        level1Items.add(createItemsForEmeraldsOffer(BWGBlocks.GREEN_MUSHROOM.get(), 1, 1, 12, 1, 0.05f));
+        level1Items.add(createItemsForEmeraldsOffer(BWGBlocks.WOOD_BLEWIT.get(), 1, 1, 12, 1, 0.05f));
+        WANDERING_TRADER_TRADES.put(1, level1Items);
     }
 
     private static MerchantOffer createEmeraldForItemsOffer(ItemLike item, int cost, int maxUses, int villagerXp) {